DEV : Simplifying process
